requirements.txt
fastapi
motor
pydantic
python-jose
passlib
bcrypt
uvicorn
python-dotenv